successful point distribution

+++ index.php
@@ -2,10 +2,11 @@
 session_start();
 include "./class/database.php";
 include "./class/event.php";
-include "./class/user.php";
+include "./class/user_event.php";
 
 $events = new Event();
 $eventData = $events->getAllEvents();
+$user_event = new UserEvent();
 ?>
 
 <!DOCTYPE html>
@@ -89,36 +90,39 @@ $eventData = $events->getAllEvents();
                 <div><button onclick="location.href='eventlist.php?sort=help'">協辦</button></div>
                 <div><button onclick="location.href='eventlist.php?sort=attend'">參加</button></div>
             </div>
-            <div id="eventlist">         
+            <div id="eventlist">  
                     <?php foreach($eventData as $event):?>
-                        <div class="events">
-                            <a id='imga'href="./eventmore.php?id=<?= $event["id"]?>">
-                                <div class="eventimg" id="coverpic"
-                                style="background-image: url('<?= $event["image"]?>');" ></div>
-                            </a>
-                            <div>
-                                <h3 id="name"><?= $event["name"]?></h3>
-                                <p id="time" href=""><?= $event["date"]?>   <?= $event["time"]?></p>
-                                <a id="location" href="https://www.google.com.tw/maps/search/<?= $event["location"]?>"><?= $event["location"]?></a><br>
-                                
-                            </div>
-                            <?php if(isset($_SESSION["user_id"])):?>
-                                <?php if($_SESSION["user_id"] != $event["holder"]):?>
-                                    <div id="btnarea">
-                                        <div id="interested" onclick="location.href='favorite.php?id=<?= $event["id"]?>';"></div>
-                                        <div onclick="alert('請先選擇協辦或參加')">attend</div>
-                                    </div>
+                        <?php $user_event->event_id = $event["id"]?>
+                        <?php if($user_event->getStatus() == 'not_yet'):?>
+                            <div class="events">
+                                <a id='imga'href="./eventmore.php?id=<?= $event["id"]?>">
+                                    <div class="eventimg" id="coverpic"
+                                    style="background-image: url('<?= $event["image"]?>');" ></div>
+                                </a>
+                                <div>
+                                    <h3 id="name"><?= $event["name"]?></h3>
+                                    <p id="time" href=""><?= $event["date"]?>   <?= $event["time"]?></p>
+                                    <a id="location" href="https://www.google.com.tw/maps/search/<?= $event["location"]?>"><?= $event["location"]?></a><br>
+                                    
+                                </div>
+                                <?php if(isset($_SESSION["user_id"])):?>
+                                    <?php if($_SESSION["user_id"] != $event["holder"]):?>
+                                        <div id="btnarea">
+                                            <div id="interested" onclick="location.href='favorite.php?id=<?= $event["id"]?>';"></div>
+                                            <div onclick="alert('請先選擇協辦或參加')">attend</div>
+                                        </div>
+                                    <?php else:?>
+                                        <div id="btnarea">
+                                            <div id="edit">編輯</div>
+                                            <div id="delete" onclick="location.href='delete.php?id=<?= $event["id"]?>';">刪除</div>
+                                        </div>
+                                    <?php endif;?>
                                 <?php else:?>
                                     <div id="btnarea">
-                                        <div id="edit">編輯</div>
-                                        <div id="delete" onclick="location.href='delete.php?id=<?= $event["id"]?>';">刪除</div>
                                     </div>
                                 <?php endif;?>
-                            <?php else:?>
-                                <div id="btnarea">
-                                </div>
-                            <?php endif;?>
-                        </div>
+                            </div>
+                        <?php endif;?>
                     <?php endforeach;?>
             </div>
